requirements.txt
fastapi
pytest
requests
uvicorn
flake8
mypy
pydantic-settings
sqlalchemy
black
python-jose
python-dotenv
sqlmodel
psycopg2
rich    
pyjwt
ptpython
asyncpg
passlib==1.7.4 
bcrypt==3.2.2
redis==4.0.0